maintain waybar profile on reload/reboot

+++ hypr/scripts/waybar_profile.sh
@@ -3,14 +3,22 @@
 menu=("Topbar" "Bottombar")
 choice=$(printf "%s\n" "${menu[@]}" | fuzzel --dmenu -i -p "Waybar Profile: ")
 
+if [ -z "$choice" ]; then
+    exit 1
+fi
+
 waybar_pid=$(pgrep -x waybar)
-kill $waybar_pid
+if [ -n "$waybar_pid" ]; then
+  kill $waybar_pid
+fi
 
 case $choice in
     Topbar)
-      hyprctl dispatch exec "waybar -c $HOME/.config/waybar/topbar/config.jsonc -s $HOME/.config/waybar/topbar/style.css"
+      bar="topbar"
       ;;
     Bottombar)
-      hyprctl dispatch exec "waybar -c $HOME/.config/waybar/bottombar/config.jsonc -s $HOME/.config/waybar/bottombar/style.css"
+      bar="bottombar"
       ;;
 esac
+
+echo "exec = pkill waybar; waybar -c $HOME/.config/waybar/$bar/config.jsonc -s $HOME/.config/waybar/$bar/style.css" > $HOME/.config/hypr/autostart/bar_profile.conf
